requirements.txt
aioconsole==0.8.2
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.22.1
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
bech32==1.2.0
cachetools==6.2.4
cffi==2.0.0
cryptography==46.0.3
frozenlist==1.8.0
idna==3.11
liboqs-python @ file:///Users/trbouma/projects/liboqs-python
monstr==0.1.9
multidict==6.7.0
propcache==0.4.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
secp256k1==0.14.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
yarl==1.22.0
